% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTable_utils.R
\name{build_group_vec}
\alias{build_group_vec}
\title{build_group_vec}
\usage{
build_group_vec(DATA, GRP_PATTERNS)
}
\arguments{
\item{DATA}{Dataframe, with rows of features and columns of samples}

\item{GRP_PATTERNS}{Patterns to identify group columns}
}
\value{
GRP_VEC Vector of group names
}
\description{
generates vector of group names from sample names
}
