% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTable_Normalization.R
\name{get_groupRLA}
\alias{get_groupRLA}
\title{get_groupRLA}
\usage{
get_groupRLA(DATA, GRP_PATTERNS)
}
\arguments{
\item{DATA}{Dataframe, where rows are mass feaures and columns are samples}

\item{GRP_PATTERNS}{Patterns to identify sample groups}
}
\value{
Dataframe with relative log abundances of mass features
}
\description{
Get relative log abundances of mass features compared to mass feature median.
}
