% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTable_Denoising.R
\name{get_mf_blankInfo}
\alias{get_mf_blankInfo}
\title{get_mf_blankInfo}
\usage{
get_mf_blankInfo(MF_DF, BLANK_PATTERN)
}
\arguments{
\item{MF_DF}{Dataframe, where rows are mass feaures and columns are samples}

\item{BLANK_PATTERN}{Pattern to identify blank samples}
}
\value{
Dataframe with mass feature information from blanks
}
\description{
Get information on mass features from blanks. Uses BLANK_PATTERN to identify blank columns. Unmatched columns are interpreted as sample columns.
}
