% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTable_Denoising.R
\name{get_mf_noiseInfo}
\alias{get_mf_noiseInfo}
\title{get_mf_noiseInfo}
\usage{
get_mf_noiseInfo(MF_DF, GRP_PATTERNS)
}
\arguments{
\item{MF_DF}{Mass feature dataframe, where rows are mass feaures and columns are samples}

\item{GRP_PATTERNS}{Patterns to identify sample groups}
}
\value{
Dataframe with information on mass feature noisiness.
}
\description{
Determine the prevalence of each mass feature in sample groups. If mass feature is not conserved in any sample group (detected in all replicates of sample group) then it is considered noisy. If mass feature is missing in any sample group replicates then it's considered group noisy.
}
