% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTable_Normalization.R
\name{rowwise_pw_test}
\alias{rowwise_pw_test}
\title{rowwise_pw_test}
\usage{
rowwise_pw_test(DATA, GRP_VEC, TEST_NAME, PAIRED = False)
}
\arguments{
\item{DATA}{Dataframe, where rows are mass feaures and columns are samples}

\item{GRP_VEC}{Vector of sample group names}

\item{TEST_NAME}{String of R function to perform pairwise test (e.g. t.test, kw.test)}

\item{PAIRED}{Boolean, if paired test then set to True (Default = False)}
}
\value{
Dataframe with relative log abundances of mass features
}
\description{
Perform pairwise statistical test between groups.
}
